Use set operations for argument checking
