+++ tests/fixtures.js
@@ -1,0 +1,256 @@
+exports.sensorTree = {
+    array_of_things_chicago: {
+    "0000001e0610b9e7": {
+        tsys01: { temperature: "temperature.temperature" },
+        "apds-9006-020": { intensity: "light_intensity.500nm" },
+        chemsense: {
+        co: "gas_concentration.co",
+        reducing_gases: "gas_concentration.reducing_gases",
+        h2s: "gas_concentration.h2s",
+        so2: "gas_concentration.so2",
+        oxidizing_gases: "gas_concentration.oxidizing_gases",
+        o3: "gas_concentration.o3",
+        no2: "gas_concentration.no2"
+        },
+        mma8452q: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        ml8511: { intensity: "light_intensity.365nm" },
+        tsl260rd: { intensity: "light_intensity.940nm" },
+        hih6130: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        bmp180: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        sht25: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        hih4030: { humidity: "relative_humidity.humidity" },
+        bmi160: {
+        accel_z: "acceleration.z",
+        accel_x: "acceleration.x",
+        accel_y: "acceleration.y",
+        orient_z: "orientation.z",
+        orient_x: "orientation.x",
+        orient_y: "orientation.y"
+        },
+        htu21d: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        lps25h: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        tmp421: { temperature: "temperature.internal_temperature" },
+        tmp112: { temperature: "temperature.temperature" },
+        hmc5883l: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        mlx75305: { intensity: "light_intensity.700nm" },
+        pr103j2: { temperature: "temperature.temperature" }
+    },
+    "0000001e0610ba89": {
+        tsys01: { temperature: "temperature.temperature" },
+        "apds-9006-020": { intensity: "light_intensity.500nm" },
+        chemsense: {
+        co: "gas_concentration.co",
+        reducing_gases: "gas_concentration.reducing_gases",
+        h2s: "gas_concentration.h2s",
+        so2: "gas_concentration.so2",
+        oxidizing_gases: "gas_concentration.oxidizing_gases",
+        o3: "gas_concentration.o3",
+        no2: "gas_concentration.no2"
+        },
+        mma8452q: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        ml8511: { intensity: "light_intensity.365nm" },
+        tsl260rd: { intensity: "light_intensity.940nm" },
+        hih6130: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        bmp180: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        sht25: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        hih4030: { humidity: "relative_humidity.humidity" },
+        bmi160: {
+        accel_z: "acceleration.z",
+        accel_x: "acceleration.x",
+        accel_y: "acceleration.y",
+        orient_z: "orientation.z",
+        orient_x: "orientation.x",
+        orient_y: "orientation.y"
+        },
+        htu21d: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        lps25h: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        tmp421: { temperature: "temperature.internal_temperature" },
+        tmp112: { temperature: "temperature.temperature" },
+        hmc5883l: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        mlx75305: { intensity: "light_intensity.700nm" },
+        pr103j2: { temperature: "temperature.temperature" }
+    },
+    "0000001e0610b9fd": {
+        tsys01: { temperature: "temperature.temperature" },
+        "apds-9006-020": { intensity: "light_intensity.500nm" },
+        chemsense: {
+        co: "gas_concentration.co",
+        reducing_gases: "gas_concentration.reducing_gases",
+        h2s: "gas_concentration.h2s",
+        so2: "gas_concentration.so2",
+        oxidizing_gases: "gas_concentration.oxidizing_gases",
+        o3: "gas_concentration.o3",
+        no2: "gas_concentration.no2"
+        },
+        mma8452q: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        ml8511: { intensity: "light_intensity.365nm" },
+        tsl260rd: { intensity: "light_intensity.940nm" },
+        hih6130: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        bmp180: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        sht25: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        hih4030: { humidity: "relative_humidity.humidity" },
+        bmi160: {
+        accel_z: "acceleration.z",
+        accel_x: "acceleration.x",
+        accel_y: "acceleration.y",
+        orient_z: "orientation.z",
+        orient_x: "orientation.x",
+        orient_y: "orientation.y"
+        },
+        htu21d: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        lps25h: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        tmp421: { temperature: "temperature.internal_temperature" },
+        tmp112: { temperature: "temperature.temperature" },
+        hmc5883l: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        mlx75305: { intensity: "light_intensity.700nm" },
+        pr103j2: { temperature: "temperature.temperature" }
+    },
+    "0000001e0610ba72": {
+        tsys01: { temperature: "temperature.temperature" },
+        "apds-9006-020": { intensity: "light_intensity.500nm" },
+        chemsense: {
+        co: "gas_concentration.co",
+        reducing_gases: "gas_concentration.reducing_gases",
+        h2s: "gas_concentration.h2s",
+        so2: "gas_concentration.so2",
+        oxidizing_gases: "gas_concentration.oxidizing_gases",
+        o3: "gas_concentration.o3",
+        no2: "gas_concentration.no2"
+        },
+        mma8452q: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        ml8511: { intensity: "light_intensity.365nm" },
+        tsl260rd: { intensity: "light_intensity.940nm" },
+        hih6130: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        bmp180: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        sht25: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        hih4030: { humidity: "relative_humidity.humidity" },
+        bmi160: {
+        accel_z: "acceleration.z",
+        accel_x: "acceleration.x",
+        accel_y: "acceleration.y",
+        orient_z: "orientation.z",
+        orient_x: "orientation.x",
+        orient_y: "orientation.y"
+        },
+        htu21d: {
+        temperature: "temperature.temperature",
+        humidity: "relative_humidity.humidity"
+        },
+        lps25h: {
+        pressure: "atmospheric_pressure.pressure",
+        temperature: "temperature.temperature"
+        },
+        tmp421: { temperature: "temperature.internal_temperature" },
+        tmp112: { temperature: "temperature.temperature" },
+        hmc5883l: {
+        y: "magnetic_field.y",
+        x: "magnetic_field.x",
+        z: "magnetic_field.z"
+        },
+        mlx75305: { intensity: "light_intensity.700nm" },
+        pr103j2: { temperature: "temperature.temperature" }
+    }
+    },
+    plenario_development: {
+    node_dev_2: {
+        sensor_dev_3: {
+        n2: "gas_concentration.n2",
+        co2: "gas_concentration.co2"
+        },
+        "sensor_dev_2 ": { humidity: "relative_humidity.humidity" }
+    },
+    node_dev_1: {
+        sensor_dev_1: {
+        mag_y: "magnetic_field.y",
+        mag_x: "magnetic_field.x"
+        },
+        sensor_dev_4: {
+        oxygen: "gas_concentration.o2",
+        temp: "temperature.temperature",
+        mag_z: "magnetic_field.z"
+        }
+    }
+    }
+};
